requirements.txt
pandas
numpy
requests
yfinance
ib_insync
pyyaml
scipy
pandas_ta
matplotlib
joblib
tqdm